Modularise Effect and Effect.Internal better
